# Enum validator improvements