## Change Summary

While looking through how we do enum validation in preparation for moving some of it to Rust, I saw a few things that could be improved. Hopefully this provides a "cleaner" schema, and should improve performance a bit.

Performance gain is small, but not insignificant:

### before

```py
In [1]: import enum
   ...: from pydantic import TypeAdapter
   ...: 
   ...: class Foo(enum.IntEnum):
   ...:     x = 1
   ...:     y = 2
   ...: 
   ...: 
   ...: ta = TypeAdapter(Foo)
   ...: 
   ...: %timeit ta.validate_python(2)
   ...: %timeit ta.validate_json('2')
   ...: 
550 ns ± 2.37 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
560 ns ± 2.29 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

### After

```py
In [1]: import enum
   ...: from pydantic import TypeAdapter
   ...: 
   ...: class Foo(enum.IntEnum):
   ...:     x = 1
   ...:     y = 2
   ...: 
   ...: 
   ...: ta = TypeAdapter(Foo)
   ...: 
   ...: %timeit ta.validate_python(2)
   ...: %timeit ta.validate_json('2')
   ...: 
508 ns ± 0.84 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
517 ns ± 1.95 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

## Checklist

* [x] The pull request title is a good summary of the changes - it will be used in the changelog
* [x] Unit tests for the changes exist
* [ ] Tests pass on CI
* [ ] Documentation reflects the changes where applicable
* [ ] My PR is ready to review, **please add a comment including the phrase "please review" to assign reviewers**